requirements.txt
Flask
ics
Flask-HTTPAuth
psycopg2-binary